add result type and use that to handle parsing errors pass embedded object instead of attachment info to embedded object constructors
At the end of the day, try/catch vs a functional-style result type
error handling strategy is a purely stylistic choice and doesnt really
matter. At least, in my opinion. I just wanted to play around with
making a Result-ish type in Java and try it out to see how it feels to
use it to handle errors instead of the more traditional try/catch strategy.